Safari-611.1.10/LayoutTests/platform/wpe/imported/w3c/web-platform-tests/pointerevents/pointerlock/pointerevent_pointermove_in_pointerlock-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: target0.requestPointerLock is not a function. (In 'target0.requestPointerLock()', 'target0.requestPointerLock' is undefined)
pointerlock

Test Description: This test checks if pointermove is dispatched correctly while in pointerlock mode.
Click in the black box (and accept the pointerlock permission if asked).
Move your mouse.
Click in the purple box inside the iframe
Move your mouse.



Harness Error (FAIL), message = Unhandled rejection: target0.requestPointerLock is not a function. (In 'target0.requestPointerLock()', 'target0.requestPointerLock' is undefined)

NOTRUN pointermove event received
NOTRUN pointermove event received inner frame
